pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "data_product_tracker"
version = "1.0.0-beta.1"  # Managed by python-semantic-release
description = "An interface to track various data products and their dependencies and environments"
authors = [{name="William Fong", email="willfong@mit.edu"}]
readme = "README.rst"

dependencies = [
    "sqlalchemy>=2.0",
    "click>=8.0",
    "setuptools>69.5",
    "deal",
    "mmh3",
    "psycopg[binary]",
    "kavli-configurables",
]

[project.urls]
Home = "https://tessgit.mit.edu/wcfong/data-product-tracker"


[project.optional-dependencies]
dev = [
    "black>=24.4",
    "hypothesis>=6",
    "hypothesis_fspaths",
    "mypy==1.10",
    "pytest-cov",
    "pytest-mock",
    "pytest-sugar",
    "pytest-xdist",
    "pytest>=8.2",
    "nox>=2023.4.22",
    "python-semantic-release>=8.0",
    "build>=1.0",
    "flake8>=6.0",
    "flake8-docstrings>=1.7",
    "flake8-bugbear>=23.0",
    "isort>=5.12",
    "sphinx>=7.0",
    "sphinx-rtd-theme>=1.3",
    "coverage[toml]>=7.0",
    "safety>=2.3",
]

[tool.mypy]
disable_error_code = "import-untyped"
plugins = "sqlalchemy.ext.mypy.plugin"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.flake8]
max-line-length = 81

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]
pythonpath = ["src"]
# Configure pytest-xdist for SQLite compatibility
markers = [
    "serial: marks tests that should run serially (not in parallel)",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variable = ["src/data_product_tracker/__init__.py:__version__"]
branch = "main"
upload_to_pypi = false
upload_to_release = true
build_command = "pip install build && python -m build"
commit_parser = "conventional"
major_on_zero = false
tag_format = "v{version}"

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.branches.staging]
match = "staging"
prerelease = true
prerelease_token = "beta"

[tool.semantic_release.commit_parser_options]
allowed_types = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_types = ["feat"]
patch_types = ["fix", "perf", "refactor", "build", "chore", "ci", "docs", "style", "test"]
